requirements/base.txt
Django==3.2.4
djangorestframework==3.14.0
django-fitler==22.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1

